Login/auth, payments table, booking→leads.pre_*, settings tabs
- Auth: scrypt login, x-user-id header, /api/login, auth guard, logout
- Settings tabs: General + Users (admin CRUD)
- payments table with step_no/doc_no + atomic confirm in PaymentSection
- Move bookings table into leads.pre_* (drop LEFT JOIN)
- StepLayout, ReceiptModal/Buttons, compressImage, useSubStep
- Install signature, QR/invoice/slip verify polish

+++ src/app/invoice/[token]/page.tsx
@@ -33,7 +33,7 @@ const CO = {
 
 const fmt = (n: number) => new Intl.NumberFormat("en-US", { minimumFractionDigits: 2 }).format(n);
 const slotMap: Record<string, string> = { am: "9:00 - 12:00", morning: "9:00 - 12:00", pm: "13:00 - 16:00", afternoon: "13:00 - 16:00" };
-const thaiDate = (s: string) => new Date(s.slice(0, 10) + "T12:00:00").toLocaleDateString("th-TH", { day: "numeric", month: "short", year: "numeric" });
+const thaiDate = (s: string) => new Date(s.slice(0, 10) + "T12:00:00").toLocaleDateString("th-TH", { day: "numeric", month: "long", year: "numeric" });
 
 export default function InvoicePage() {
   const { token } = useParams();
